Refactoring: move test errors out of test runners folder
